text formatters to ts which implement an interface. adjust text buffer to deal with this
