+++ javascript/src/TextBuffer/TextFormatter.ts
@@ -1,0 +1,3 @@
+interface TextFormatter {
+    format(strArr: Array<string>, ctx?: {}): Array<string>
+}
