+++ javascript/src/TextBuffer/string.extensions.ts
@@ -1,6 +1,7 @@
 interface String {
     removeCharAt(i: number): string;
     insertCharAt(i: number, c: string): string;
+    replacerec(pattern: RegExp, what: string): string;
 }
 
 String.prototype.removeCharAt = function (i: number) {
@@ -13,4 +14,11 @@ String.prototype.insertCharAt = function (i: number, c: string) {
     const l = this.slice(0, i);
     const r = this.slice(i);
     return l + c + r;
+};
+
+String.prototype.replacerec = function (pattern: RegExp, what: string) {
+    var newstr = this.replace(pattern, what);
+    if (newstr == this)
+        return newstr;
+    return newstr.replace(pattern, what);
 };
